requirements.txt
certifi==2018.8.13
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.39.1
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
EasyProcess==0.2.7
gunicorn==19.9.0
idna==2.8
lxml==4.3.4
Markdown==3.1.1
oauthlib==3.0.2
pipenv==2018.7.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
python3-openid==3.1.0
pytils==0.3
pytz==2018.7
PyVirtualDisplay==0.2.4
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
tools==0.1.9
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.3.0
whitenoise==4.1.2
